refactor: move all files to gambas directory
