+++ gambas/manager.py
