+++ gambas/run.py
@@ -1,0 +1,5 @@
+from gambas.command import Command
+
+def main():
+  cmd = Command()
+  cmd.run()
