+++ gambas/tests/test_validator.py
@@ -1,6 +1,6 @@
 import pytest
-from tests.conftest import TEST_DEFAULT_CONFIG_DICT, BASE_DIR 
-from validator import ConfigValidator
+from gambas.tests.conftest import TEST_DEFAULT_CONFIG_DICT, BASE_DIR 
+from gambas.validator import ConfigValidator
 
 
 class TestValidator:
