--- run.py
@@ -1,5 +1,0 @@
-from command import Command
-
-if __name__ == "__main__":
-  cmd = Command()
-  cmd.run()
